Restructure more similar to workshops.
